timetable_generator/requirements.txt
Django>=5.2.6
djangorestframework>=3.16.1
psycopg2-binary>=2.9.10
celery>=5.4.0
redis>=5.0.0
django-cors-headers>=4.3.1